An initial attempt to get a remote VNC display working.
It's got a lot of problems, but it works.

Password authentication works.

Still need a "forceAuthScheme" property because auth scheme negotiation doesn't work right;
that seems to be in the VNC client; maybe using a more recent version of noVNC would fix this

another problem with this code - the react component is unmaintained, and well behind its upstream

add a new action bar item for sharing remote desktop

The remote desktop URL is entered into a dialog box, but
the password is hard-wired, and it cuts off the
bottom of the remote desktop screen instead of resizing it.
   (that's why the style sheet is changed from 'center' to 'stretch'; otherwise both top and bottom get cut off)