move code that sets window.VncDisplay into the "ref" block
This change, while more logical, means that window.VncDisplay can be
invalid in two different ways - a remote desktop was never shared
(so it's undefined), or a remote desktop was shared and then
unshared (so it's null).

The ultimate goal is to depricate this code - storing a copy of
the VNC display in the global 'window' dictionary isn't a very
good solution to letting the lock button access the desktop component.